move client rehydration types to types file
