+++ src/channel_state.ts
@@ -2,6 +2,8 @@ import { Channel } from './channel';
 import {
   ChannelMemberResponse,
   ChannelMembership,
+  ChannelStateDataInput,
+  ChannelStateDataOutput,
   FormatMessageResponse,
   Event,
   LiteralStringForUnion,
@@ -10,105 +12,6 @@ import {
   UnknownType,
   UserResponse,
 } from './types';
-
-// Used when retrieving the state data
-export type ChannelStateDataOutput<
-  AttachmentType extends UnknownType = UnknownType,
-  ChannelType extends UnknownType = UnknownType,
-  CommandType extends string = LiteralStringForUnion,
-  MessageType extends UnknownType = UnknownType,
-  ReactionType extends UnknownType = UnknownType,
-  UserType extends UnknownType = UnknownType
-> = {
-  isUpToDate: boolean;
-  last_message_at: Date | null;
-  members: Record<string, ChannelMemberResponse<UserType>>;
-  membership: ChannelMembership<UserType>;
-  messages: FormatMessageResponse<
-    AttachmentType,
-    ChannelType,
-    CommandType,
-    MessageType,
-    ReactionType,
-    UserType
-  >[];
-  mutedUsers: Array<UserResponse<UserType>>;
-  pinnedMessages: FormatMessageResponse<
-    AttachmentType,
-    ChannelType,
-    CommandType,
-    MessageType,
-    ReactionType,
-    UserType
-  >[];
-  read: Record<string, { last_read: Date; user: UserResponse<UserType> }>;
-  threads: Record<
-    string,
-    Array<
-      FormatMessageResponse<
-        AttachmentType,
-        ChannelType,
-        CommandType,
-        MessageType,
-        ReactionType,
-        UserType
-      >
-    >
-  >;
-  unreadCount: number;
-};
-
-// Used when reInitializing the state data
-export type ChannelStateDataInput<
-  AttachmentType extends UnknownType = UnknownType,
-  ChannelType extends UnknownType = UnknownType,
-  CommandType extends string = LiteralStringForUnion,
-  MessageType extends UnknownType = UnknownType,
-  ReactionType extends UnknownType = UnknownType,
-  UserType extends UnknownType = UnknownType
-> = Omit<
-  ChannelStateDataOutput<
-    AttachmentType,
-    ChannelType,
-    CommandType,
-    MessageType,
-    ReactionType,
-    UserType
-  >,
-  'last_message_at' | 'messages' | 'pinnedMessages' | 'read' | 'threads'
-> & {
-  last_message_at: string | null;
-  messages: MessageResponse<
-    AttachmentType,
-    ChannelType,
-    CommandType,
-    MessageType,
-    ReactionType,
-    UserType
-  >[];
-  pinnedMessages: MessageResponse<
-    AttachmentType,
-    ChannelType,
-    CommandType,
-    MessageType,
-    ReactionType,
-    UserType
-  >[];
-  read: Record<string, { last_read: string; user: UserResponse<UserType> }>;
-  threads: Record<
-    string,
-    Array<
-      MessageResponse<
-        AttachmentType,
-        ChannelType,
-        CommandType,
-        MessageType,
-        ReactionType,
-        UserType
-      >
-    >
-  >;
-};
 
 /**
  * ChannelState - A container class for the channel state.
